add assembly and deepgram model selection
