+++ src/llms/fireworks.ts
@@ -1,9 +1,9 @@
 // src/llms/fireworks.ts
 
 import { env } from 'node:process'
-import { FIREWORKS_MODELS } from '../utils/llm-globals'
+import { FIREWORKS_MODELS } from '../utils/globals/llms'
 import { err, logAPIResults } from '../utils/logging'
-import type { LLMFunction, FireworksModelType, FireworksResponse } from '../types/llms'
+import type { LLMFunction, FireworksModelType, FireworksResponse } from '../utils/types/llms'
 
 /**
  * Main function to call Fireworks AI API.
