+++ src/process-commands/playlist.ts
@@ -7,11 +7,11 @@
 
 import { processVideo } from './video'
 import { savePlaylistInfo } from '../utils/save-info'
-import { execFilePromise } from '../utils/globals'
+import { execFilePromise } from '../utils/globals/process'
 import { l, err, logPlaylistSeparator } from '../utils/logging'
-import type { ProcessingOptions } from '../types/process'
-import type { TranscriptServices } from '../types/transcription'
-import type { LLMServices } from '../types/llms'
+import type { ProcessingOptions } from '../utils/types/process'
+import type { TranscriptServices } from '../utils/types/transcription'
+import type { LLMServices } from '../utils/types/llms'
 
 /**
  * Processes an entire YouTube playlist by:
