+++ src/process-steps/03-run-transcription.ts
@@ -4,8 +4,8 @@ import { callWhisper } from '../transcription/whisper'
 import { callDeepgram } from '../transcription/deepgram'
 import { callAssembly } from '../transcription/assembly'
 import { l, err } from '../utils/logging'
-import type { ProcessingOptions } from '../types/process'
-import type { TranscriptServices } from '../types/transcription'
+import type { ProcessingOptions } from '../utils/types/process'
+import type { TranscriptServices } from '../utils/types/transcription'
 
 /**
  * Orchestrates the transcription process using the specified service.
@@ -31,13 +31,21 @@ export async function runTranscription(
   try {
     switch (transcriptServices) {
       case 'deepgram':
-        const deepgramTranscript = await callDeepgram(options, finalPath)
+        // If user typed `--deepgram BASE`, then `options.deepgram` will be "BASE"
+        // If user typed just `--deepgram`, then `options.deepgram` will be true
+        const deepgramModel = typeof options.deepgram === 'string' ? options.deepgram : 'NOVA_2'
+        const deepgramTranscript = await callDeepgram(options, finalPath, deepgramModel)
         l.wait('\n  Deepgram transcription completed successfully.\n')
+        l.wait(`\n    - deepgramModel: ${deepgramModel}`)
         return deepgramTranscript
 
       case 'assembly':
-        const assemblyTranscript = await callAssembly(options, finalPath)
+        // If user typed `--assembly NANO`, then `options.assembly` will be "NANO"
+        // If user typed just `--assembly`, then `options.assembly` will be true
+        const assemblyModel = typeof options.assembly === 'string' ? options.assembly : 'NANO'
+        const assemblyTranscript = await callAssembly(options, finalPath, assemblyModel)
         l.wait('\n  AssemblyAI transcription completed successfully.\n')
+        l.wait(`\n    - assemblyModel: ${assemblyModel}`)
         return assemblyTranscript
 
       case 'whisper':
